:sparkles: feat(S3) : File entity의 필요성이 없다고, 판단 관련 삭제
