--- src/main/java/com/ip/ddangddangddang/domain/auction/repository/FileRepository.java
@@ -1,10 +1,0 @@
-package com.ip.ddangddangddang.domain.auction.repository;
-
-import com.ip.ddangddangddang.domain.auction.entity.File;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FileRepository extends JpaRepository<File, Long> {
-
-}
